app/requirements.txt
blinker==1.6.2
click==8.1.5
Flask==2.3.3
Flask-Session==0.8.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
zipp==3.16.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7
requests==2.26.0
google_auth_oauthlib==0.4.6
google_auth==2.28.1
oauth2client==4.1.3